Fix #817
GetEntropyRandomSeed.h incudes <sys/syscall.h> which is not available in
Windows enviroment.

By using _WIN32 preprocessor definition we disable this
inclusion from happening in Windows based systems. _WIN32 is defined
in cygwin and mingw enviroments.

It is not possible, from the preprocessor, to check for header file.
It could be done from buildscript but current solution works as well.

Conflicts:
	node_build/dependencies/libuv