Fix v2 long outgoing references blocking cjdroute
version 2 of this fix (styling and also detect NULL
password and public key)

Too long text of reference in outgoing peers
(in connectTo) is handled correctly:

We send to cjdroute (via UDP packet) only
the needed data, skipping the fields
that are not used.

Also, if anyway the reference is still too large
(e.g. had very long password), it will be skipped
with an error logged.