Fix v3 long outgoing references blocking cjdroute
v3 we now abort if there is a broken peer.
(we leave the try-catch block in separate function,
otherwise gcc complains about clobbered variables)

Too long text of reference in outgoing peers
(in connectTo) is handled correctly:

We send to cjdroute (via UDP packet) only
the needed data, skipping the fields
that are not used.

Also, if anyway the reference is still too large
(e.g. had very long password), it will be skipped
with an error logged.